requirements.txt
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.10.0
attr==0.3.1
attrs==25.4.0
azure-core==1.36.0
azure-storage-blob==12.27.0
black==25.9.0
bleach==5.0.1
boto==2.49.0
boto3==1.40.55
botocore==1.40.55
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
datamodel-code-generator==0.26.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.13
django-annoying==0.10.6
django-cors-headers==4.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-spectacular==0.28.0
email-validator==2.3.0
expiringdict==1.2.2
Faker==37.11.0
fastapi==0.119.0
filelock==3.20.0
fonttools==4.60.1
genson==1.3.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.3
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ijson==3.4.0.post0
importlib_metadata==8.7.0
inflect==5.6.2
inflection==0.5.1
isodate==0.7.2
isort==5.13.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsf==0.11.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
label-studio==1.21.0
label-studio-sdk==2.0.4
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
loguru==0.7.3
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mypy_extensions==1.1.0
nltk==3.9.2
numpy==1.26.4
openai==1.109.1
opencv-python==4.11.0.86
opentelemetry-api==1.38.0
ordered-set==4.0.2
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
polars==1.34.0
polars-runtime-32==1.34.0
prometheus_client==0.23.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pyRFC3339==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.4
pytokens==0.2.0
pytz==2022.7.1
PyYAML==6.0.3
redis==5.2.1
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-file==3.0.0
requests-mock==1.12.1
rich==14.2.0
rpds-py==0.27.1
rq==1.16.2
rsa==4.9.1
rstr==3.2.2
rules==3.4
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
semver==3.0.4
sentry-sdk==2.42.0
setuptools==80.9.0
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.48.0
threadpoolctl==3.6.0
tldextract==5.3.0
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.37.0
watchdog==6.0.0
webencodings==0.5.1
wheel==0.40.0
wrapt==1.17.3
xmljson==0.2.1
zipp==3.23.0
